test: move parser canonicalization + legacy syntax tests to test/frontend
Refs #1152

Made-with: Cursor
