+++ test/frontend/pr252_parser_register_token_canonicalization.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest';
 
-import { parseProgram } from '../src/frontend/parser.js';
-import type { AsmInstructionNode, FuncDeclNode, OpDeclNode } from '../src/frontend/ast.js';
-import type { Diagnostic } from '../src/diagnosticTypes.js';
+import { parseProgram } from '../../src/frontend/parser.js';
+import type { AsmInstructionNode, FuncDeclNode, OpDeclNode } from '../../src/frontend/ast.js';
+import type { Diagnostic } from '../../src/diagnosticTypes.js';
 
 function asmInstructions(items: { kind: string }[]): AsmInstructionNode[] {
   return items.filter((item): item is AsmInstructionNode => item.kind === 'AsmInstruction');
